# Adding bulk insert to the Query interface