--- a/aqueduct/pubspec.yaml
+++ b/aqueduct/pubspec.yaml
@@ -5,7 +5,7 @@ homepage: https://aqueduct.io
 documentation:
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.2.0 <3.0.0"
 
 dependencies:
   args: ^1.5.0