What is contained in the pull request:

- `insertMany(List)`  method to the `Query` interface
- implementation for the `insertMany` in the postgresql implementation for Query
- tests for the new method (plus reorganization of a few tests).
- changed `insertObjects(..)` method in context to use the new method

What is the benefit:

The new implementation uses the capability to insert more then one record with a single `INSERT` query, so instead of starting a transaction and doing N inserts and then commuting. This allows for much better performance especially for larger counts of rows inserted. Here is a table with benchmark averages:

| Rows  | Old way | New way |
|-------|-----------:|----------:|
| 1     |   5.95 ms  |   4.88 ms |
| 10    |  11.47 ms  |   6.01 ms |
| 100   |  42.41 ms  |  12.93 ms |
| 1000  |  334.36 ms |  58.08 ms |
| 10000 | 3982.86 ms | 559.64 ms |